requirements.txt
# ============================================
# Gold Credit S/A - Dependencies
# ============================================

# Flask Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1

# Excel Processing
openpyxl==3.1.2
pandas==2.1.4
numpy==1.26.2
et-xmlfile==1.1.0

# Date/Time
python-dateutil==2.8.2
pytz==2024.1

# Environment Variables
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Security
email-validator==2.1.0

# Utilities
click==8.1.7
six==1.16.0
tzdata==2023.4

# Monitoramento e Saúde
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0

# Produção
whitenoise==6.6.0

# Database Migration
Flask-Migrate==4.0.5
alembic==1.13.1

# Development (optional - comment out in production)
# Flask-DebugToolbar==0.14.1
# pytest==7.4.3
# pytest-flask==1.3.0

# ============================================
# INSTALAÇÃO:
# ============================================
# 
# Desenvolvimento:
#   pip install -r requirements.txt
#
# Produção:
#   pip install --no-cache-dir -r requirements.txt
#
# Atualizar:
#   pip install --upgrade -r requirements.txt
#